Moved terminal panel out of session.
