+++ frontend/source/frontend/CMakeLists.txt
@@ -49,6 +49,7 @@ target_sources(
         session_components/file_tracking.cpp
         session_components/session_options.cpp
         session_components/operation_queue.cpp
+        session_components/terminal_panel.cpp
         sync_dialog/sync_dialog.cpp
         sync_dialog/sync_progress_dialog.cpp
         terminal/terminal_channel.cpp
